+++ src/cell.rs
@@ -1,4 +1,4 @@
-use std::{collections::BTreeSet, sync::Arc};
+use std::{collections::HashSet, sync::Arc};
 
 #[derive(
     Debug, Copy, Clone, PartialEq, Eq, Hash, PartialOrd, Ord, serde::Serialize, serde::Deserialize,
@@ -17,7 +17,7 @@ pub(crate) struct CellData {
     pub(crate) last_updated_version: u32,
     pub(crate) last_verified_version: u32,
     pub(crate) dependencies: Vec<Cell>,
-    pub(crate) input_dependencies: BTreeSet<Cell>,
+    pub(crate) input_dependencies: HashSet<Cell, rustc_hash::FxBuildHasher>,
     pub(crate) lock: Arc<parking_lot::Mutex<()>>,
 }
 
@@ -28,7 +28,7 @@ impl CellData {
             last_updated_version: 0,
             last_verified_version: 0,
             dependencies: Vec::new(),
-            input_dependencies: BTreeSet::new(),
+            input_dependencies: Default::default(),
             lock: Arc::new(parking_lot::Mutex::new(())),
         }
     }